backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
ultralytics==8.0.200
pydantic==2.5.0
python-dotenv==1.0.0
paho-mqtt==1.6.1
supabase==2.0.3
numpy==1.26.3
requests
